seed-db: re-organise fixtures, helpers
- Make clients its own fixture
- Move `populateTables` to `tests/helpers`
- Adjust seed-db script to reflect the above changes
