+++ api/tests/fixtures/cases.json
@@ -4,21 +4,7 @@
     "caseDesc": "Single family, requires education grant for son",
     "assignedAgency": "MOE",
     "agencyPoc": "chewie@moe.edu.sg",
-    "client": {
-      "name": "Han Solo",
-      "dateOfBirth": "12/12/1995",
-      "email": "han@solo.com",
-      "nric": "S9347193J",
-      "nricAddress": "123 Tatooine Street",
-      "address": "123 Tatooine Street",
-      "gender": "Male",
-      "nationality": "Jedi",
-      "race": "Caucasian",
-      "maritalStatus": "Single4Lyfe",
-      "employmentStatus": "Employed",
-      "grossHouseholdIncome": "100000",
-      "phone": 91829381
-    },
+    "clientId": 1,
     "messages": 
       [
         {
