+++ api/tests/fixtures/clients.json
@@ -1,0 +1,17 @@
+[
+  {
+    "name": "Han Solo",
+    "dateOfBirth": "12/12/1995",
+    "email": "han@solo.com",
+    "nric": "S9347193J",
+    "nricAddress": "123 Tatooine Street",
+    "address": "123 Tatooine Street",
+    "gender": "Male",
+    "nationality": "Jedi",
+    "race": "Caucasian",
+    "maritalStatus": "Single4Lyfe",
+    "employmentStatus": "Employed",
+    "grossHouseholdIncome": "100000",
+    "phone": 91829381
+  }
+]
